Moved Express application to app folder. Added json responses to controllers
