+++ app/controllers/median.js
@@ -1,43 +1,43 @@
 var _       = require("lodash");
 var pg      = require("pg");
 var utils = require('../utils');
-var conn_options = require("../rds-config.json");
+var conn_options = require("../../scripts/redshift-config.json");
 
-var quantileController = {
+var medianController = {
   process: function(req, res, next){
-    var queryParams = _.pick(req.query, ["state", "race", "sex", "agegroup", "quantile", "compare"]);
-
+    var queryParams = _.pick(req.query, ["state", "race", "sex", "agegroup", "compare"]);
     utils.validateQueryParams(queryParams, function(err, validateCallback) {
       if(err) { return next(err); }
 
       if(compare && compare !== "") {
-        quantileController.getCompareIncomeQuantiles(queryParams, res, next);
+        medianController.getCompareIncomeMedian(queryParams, res, next);
       } else {
         delete queryParams["compare"];
-        quantileController.getIncomeQuantiles(queryParams, res, next);
+        medianController.getIncomeMedian(queryParams, res, next);
       }
     });
 
   },
 
-  getIncomeQuantiles: function(queryParams, res, next){
-    var sql = "SELECT QUANTILE, INCOME FROM PUMS_2014_Quantiles";
-    sql = utils.appendWhereClause(sql, queryParams) + " ORDER BY QUANTILE::INT ASC;";
+  getIncomeMedian: function(queryParams, res, next){
+    var sql = "SELECT INCOME FROM PUMS_2014_Quantiles";
+    sql = utils.appendWhereClause(sql, queryParams) + " AND QUANTILE='50';";
+    console.log(sql)
 
     pg.connect(conn_options, function(err, client, done) {
-      if(err) { return next(err); }
+      if(err) {return next(err); }
 
       client.query(sql, function(err, response) {
         done();
-        if(err) { return next(err); }
+        if(err) { console.log(err); return next(err); }
 
         var results = response.rows;
         resultsObj = {};
         _.forEach(results, function(result) {
-          resultsObj[result["quantile"] + "%"] = result["income"];
+          resultsObj = result["income"];
         })
-
-        return next(err, {"overall": resultsObj});
+        res.json({"overall": resultsObj})
+        // return next(err, {"overall": resultsObj});
       });
     });
   },
@@ -49,8 +49,8 @@ var quantileController = {
     }
     delete queryParams["compare"];
 
-    var sql = "SELECT QUANTILE, INCOME, " + compare + " FROM PUMS_2014_Quantiles";
-    sql = utils.appendWhereClause(sql, queryParams) + " ORDER BY QUANTILE::INT ASC;";
+    var sql = "SELECT INCOME, " + compare + " FROM PUMS_2014_Quantiles";
+    sql = utils.appendWhereClause(sql, queryParams) + " AND QUANTILE='50';";
 
     pg.connect(conn_options, function(err, client, done) {
       if(err) { return next(err); }
@@ -63,10 +63,10 @@ var quantileController = {
         resultsObj = {};
         _.forEach(results, function(result) {
           if(result[compare] != "") {
-            var path = "['" + result[compare] + "']" + "['" + result["quantile"] + "%']";
+            var path = "['" + result[compare] + "']";
             _.set(resultsObj, path, result["income"]);
           }
-        })
+        });
 
         var resultsObjSorted = {};
         Object.keys(resultsObj).sort().forEach(function(key) {
@@ -79,4 +79,4 @@ var quantileController = {
   }
 }
 
-module.exports = quantileController
+module.exports = medianController
