+++ package.json
@@ -2,7 +2,7 @@
   "name": "midaas-api",
   "version": "1.0.0",
   "description": "Midaas API Express App",
-  "main": "index.js",
+  "main": "app/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
